pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "debatebench"
version = "0.1.0"
description = "CLI tool for debate-style LLM benchmarking"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Debate Bench Maintainers" }
]
dependencies = [
  "typer>=0.9,<1.0",
  "rich>=13.7,<14.0",
  "pydantic>=2.7,<3.0",
  "boto3>=1.34,<2.0",
  "PyYAML>=6.0",
  "requests>=2.31,<3.0",
  "python-dotenv>=1.0,<2.0",
  "matplotlib>=3.8,<3.9",
  "seaborn>=0.13,<0.14",
  "pandas>=2.1,<2.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<9.0"
]

[project.scripts]
debatebench = "debatebench.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["debatebench"]
